% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plant_lookup.R
\name{fungal_traits}
\alias{fungal_traits}
\alias{fungal_traits_del}
\alias{fungal_traits_version_current}
\alias{fungal_traits_versions}
\title{Fungal trait database}
\usage{
fungal_traits(version = NULL, path = NULL)

fungal_traits_versions(local = TRUE, path = NULL)

fungal_traits_version_current(local = TRUE, path = NULL)

fungal_traits_del(version, path = NULL)
}
\arguments{
\item{version}{Version number.  The default will load the most
recent version on your computer or the most recent version known
to the package if you have never downloaded the data before.
With \code{plant_lookup_del}, specifying \code{version=NULL}
will delete \emph{all} data sets.}

\item{path}{Path to store the data at.  If not given,
\code{datastorr} will use \code{rappdirs} to find the best place
to put persistent application data on your system.  You can
delete the persistent data at any time by running
\code{mydata_del(NULL)} (or \code{mydata_del(NULL, path)} if you
use a different path).}

\item{local}{Logical indicating if local or github versions should
be polled.  With any luck, \code{local=FALSE} is a superset of
\code{local=TRUE}.  For \code{mydata_version_current}, if
\code{TRUE}, but there are no local versions, then we do check
for the most recent github version.}
}
\description{
This function provides access to the fungal traits database
}
\details{
The data within this lookup table comes from many original sources.
For speedy access to the data within R use this function.  For 
access to the full meta-data, download the zip file associated with the 
Github release.
}
\examples{
#
# see the format of the resource
#
head(fungal_traits())
#
#
}

